+++ CHANGES.md
@@ -1,4 +1,4 @@
-### Apache MetaModel 4.5.1 (work-in-progress)
+### Apache MetaModel 4.5.1
 
  * [METAMODEL-227] - Fix for respecting CSV escape character also when no quote character is set.
  * [METAMODEL-183] - MongoDB module split into three: common, Mongo2 and Mongo3 to allow use of either old or new MongoDB API.